refactor: move derived token to store
