+++ src/frontend/src/eth/components/send/SendModal.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
 	import SendTokenModal from '$eth/components/send/SendTokenModal.svelte';
 	import SendTokenContext from '$eth/components/send/SendTokenContext.svelte';
-	import { token } from '$lib/derived/token.derived';
+
+	import { token } from '$lib/stores/token.store';
 </script>
 
 <SendTokenContext token={$token}>
