+++ src/frontend/src/eth/components/tokens/TokenMenu.svelte
@@ -5,10 +5,11 @@
 	import { nonNullish, notEmptyString } from '@dfinity/utils';
 	import { explorerUrl as explorerUrlStore } from '$eth/derived/network.derived';
 	import { address } from '$lib/derived/address.derived';
-	import { token, tokenStandard } from '$lib/derived/token.derived';
+	import { tokenStandard } from '$lib/derived/token.derived';
 	import type { Erc20Token } from '$eth/types/erc20';
 	import { erc20TokensInitialized } from '$eth/derived/erc20.derived';
 	import { fade } from 'svelte/transition';
+	import { token } from '$lib/stores/token.store';
 
 	let explorerUrl: string | undefined;
 	$: explorerUrl =
