+++ src/frontend/src/icp/components/convert/ConvertToBTC.svelte
@@ -7,11 +7,12 @@
 	import { waitWalletReady } from '$lib/services/actions.services';
 	import { isNullish } from '@dfinity/utils';
 	import { ckBtcMinterInfoStore } from '$icp/stores/ckbtc.store';
-	import { token, tokenId } from '$lib/derived/token.derived';
+	import { tokenId } from '$lib/derived/token.derived';
 	import type { NetworkId } from '$lib/types/network';
 	import type { OptionIcCkToken } from '$icp/types/ic';
 	import { BTC_MAINNET_NETWORK_ID } from '$env/networks.env';
 	import { i18n } from '$lib/stores/i18n.store';
+	import { token } from '$lib/stores/token.store';
 
 	const isDisabled = (): boolean =>
 		isNullish($tokenId) || isNullish($ckBtcMinterInfoStore?.[$tokenId]);
