+++ src/frontend/src/icp/components/tokens/IcTokenMenu.svelte
@@ -3,9 +3,9 @@
 	import { i18n } from '$lib/stores/i18n.store';
 	import ExternalLink from '$lib/components/ui/ExternalLink.svelte';
 	import { nonNullish } from '@dfinity/utils';
-	import { token } from '$lib/derived/token.derived';
 	import { fade } from 'svelte/transition';
 	import type { OptionIcCkToken } from '$icp/types/ic';
+	import { token } from '$lib/stores/token.store';
 
 	let explorerUrl: string | undefined;
 	$: explorerUrl = ($token as OptionIcCkToken)?.explorerUrl;
