+++ src/frontend/src/icp/components/transactions/IcTransactionLabel.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import type { Token } from '$lib/types/token';
-	import { token } from '$lib/derived/token.derived';
 	import type { OptionIcCkToken } from '$icp/types/ic';
 	import { replacePlaceholders, resolveText } from '$lib/utils/i18n.utils';
 	import { i18n } from '$lib/stores/i18n.store';
+	import { token } from '$lib/stores/token.store';
 
 	export let label: string | undefined;
 	export let fallback = '';
