+++ src/frontend/src/icp/components/transactions/IcTransactionsSkeletons.svelte
@@ -3,7 +3,8 @@
 	import SkeletonCards from '$lib/components/ui/SkeletonCards.svelte';
 	import { nonNullish } from '@dfinity/utils';
 	import { icTransactionsStore } from '$icp/stores/ic-transactions.store';
-	import { token } from '$lib/derived/token.derived';
+
+	import { token } from '$lib/stores/token.store';
 
 	let transactionsInitialized: boolean;
 	$: transactionsInitialized = nonNullish($token) && nonNullish($icTransactionsStore?.[$token.id]);
