+++ src/frontend/src/routes/(app)/transactions/+layout.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import Listener from '$lib/components/core/Listener.svelte';
-	import { token } from '$lib/derived/token.derived';
+
+	import { token } from '$lib/stores/token.store';
 </script>
 
 <Listener token={$token}>
